+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@ Version 3.34.0-eisop2 (May ?, 2023)
 
 **User-visible changes:**
 
+The Initialization Checker supports new qualifier:
+* `@PolyInitialized`
+
 **Implementation details:**
 
 Deprecated `ObjectCreationNode#getConstructor` in favor of new `ObjectCreationNode#getTypeToInstantiate()`.
@@ -12,7 +15,7 @@ The class was deprecated in release 3.21.3-eisop1 (March 23, 2022) and no longer
 
 **Closed issues:**
 
-eisop#376.
+eisop#376, eisop#494
 
 
 Version 3.34.0-eisop1 (May 9, 2023)